Refactor legacy FASTA import (ready for NCBI import)
